Finished a EditProfileActivity, however it is undocumented/not documented properly.
Added some vector assets for cleaner UI.

While everything seems to work. I will need to change the way we upload, and use images since we need to pay for a better version of Firebase for this work.

Currently prioritizing figuring out how to convert an image as Base64 in Realtime Database. This is needd so we can save the string to our free DB, then load it back into a bitmap like we done in this version.